// types.ts
export type ArticleCategory = 'sea-freight' | 'customs' | 'ports' | 'news' | 'logistics-tech' | 'warehousing' | 'global-trade' | 'financial' | 'educational' | 'laws-and-regulations';
// ...
export interface Article {
  id: number;
  category: ArticleCategory;
  title: {
    ar: string;
    en: string;
  };
  summary: {
    ar: string;
    en: string;
  };
  content: {
    ar: string;
    en: string;
  };
  imageUrl: string;
  date?: {
    ar: string;
    en: string;
  };
}